setup changing mapper to presenter
