+++ spec/csv_row_model/import/presenter/concern_spec.rb
@@ -1,0 +1,4 @@
+require 'spec_helper'
+
+describe CsvRowModel::Import::Presenter::Concern do
+end
